audio part seperated from video editor
